backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-chess==1.999
torch==2.1.1
numpy==1.26.2
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
supabase==2.0.3 